fix: change BuildVersion to int64 allowing larger build numbers

+++ step/models.go
@@ -5,8 +5,8 @@ type Input struct {
 	Scheme                  string `env:"scheme,required"`
 	Target                  string `env:"target"`
 	Configuration           string `env:"configuration"`
-	BuildVersion            int    `env:"build_version,required"`
-	BuildVersionOffset      int    `env:"build_version_offset"`
+	BuildVersion            int64  `env:"build_version,required"`
+	BuildVersionOffset      int64  `env:"build_version_offset"`
 	BuildShortVersionString string `env:"build_short_version_string"`
 }
 
@@ -15,11 +15,11 @@ type Config struct {
 	Scheme                  string
 	Target                  string
 	Configuration           string
-	BuildVersion            int
-	BuildVersionOffset      int
+	BuildVersion            int64
+	BuildVersionOffset      int64
 	BuildShortVersionString string
 }
 
 type Result struct {
-	BuildVersion int
+	BuildVersion int64
 }
